+++ backend/src/main/java/com/grepp/smartwatcha/app/controller/web/admin/movie/upcoming/UpcomingMoviePageController.java
@@ -1,6 +1,6 @@
 package com.grepp.smartwatcha.app.controller.web.admin.movie.upcoming;
 
-import com.grepp.smartwatcha.app.model.admin.movie.upcoming.repository.jpa.UpcomingMovieSyncTimeJpaRepository;
+import com.grepp.smartwatcha.app.model.admin.movie.upcoming.repository.UpcomingMovieSyncTimeJpaRepository;
 import com.grepp.smartwatcha.app.model.admin.movie.upcoming.service.jpa.UpcomingMovieSaveJpaService;
 import com.grepp.smartwatcha.infra.jpa.entity.MovieEntity;
 import com.grepp.smartwatcha.infra.jpa.entity.SyncTimeEntity;
@@ -20,10 +20,10 @@ import org.springframework.web.bind.annotation.RequestParam;
 /*
  * 공개 예정작 관리 페이지 컨트롤러
  * DB 에서 공개 예정작을 조회하여 관리자 페이지에 표시
- * 
+ *
  * 입력: 페이지 번호, 페이지 크기
  * 출력: 공개 예정작 목록 페이지 (movie.html)
- * 
+ *
  * 기능:
  * - 공개 예정작 목록 조회 (개봉일 기준 오름차순 정렬)
  * - 페이지네이션 처리
@@ -39,16 +39,16 @@ public class UpcomingMoviePageController {
 
     /*
      * 공개 예정작 목록 페이지 조회
-     * 
+     *
      * 입력:
      * - page: 페이지 번호 (기본값: 0)
      * - size: 페이지당 항목 수 (기본값: 5)
-     * 
+     *
      * 출력:
      * - movies: 공개 예정작 목록
      * - pageResponse: 페이지네이션 정보
      * - lastSyncTime: 마지막 동기화 시간
-     * 
+     *
      * 정렬:
      * - 개봉일 기준 오름차순 정렬 (최근 개봉작이 아래에 표시)
      */
