+++ backend/src/main/java/com/grepp/smartwatcha/app/model/admin/AdminDashboardJpaService.java
@@ -1,7 +1,7 @@
 package com.grepp.smartwatcha.app.model.admin;
 
 import com.grepp.smartwatcha.app.model.admin.movie.list.repository.AdminMovieJpaRepository;
-import com.grepp.smartwatcha.app.model.admin.movie.upcoming.repository.jpa.UpcomingMovieSyncTimeJpaRepository;
+import com.grepp.smartwatcha.app.model.admin.movie.upcoming.repository.UpcomingMovieSyncTimeJpaRepository;
 import com.grepp.smartwatcha.app.model.admin.tag.AdminTagJpaRepository;
 import com.grepp.smartwatcha.app.model.admin.user.repository.AdminUserJpaRepository;
 import com.grepp.smartwatcha.infra.jpa.entity.MovieEntity;
